+++ xenoblade_blender/import_wismhd.py
@@ -32,17 +32,6 @@ class ImportWismhd(bpy.types.Operator, ImportHelper):
         maxlen=255,
     )
 
-    game_version: EnumProperty(
-        name="Game Version",
-        description="The game version for the shader database",
-        items=(
-            ("XC1", "Xenoblade 1 DE", "Xenoblade Chronicles 1 Definitive Edition"),
-            ("XC2", "Xenoblade 2", "Xenoblade Chronicles 2"),
-            ("XC3", "Xenoblade 3", "Xenoblade Chronicles 3"),
-        ),
-        default="XC3",
-    )
-
     pack_images: BoolProperty(
         name="Pack Images",
         description="Pack all images into the Blender file. Increases memory usage and import times but makes the Blender file easier to share by not creating additional files",
@@ -66,7 +55,7 @@ class ImportWismhd(bpy.types.Operator, ImportHelper):
         log_fmt = "%(levelname)s %(name)s %(filename)s:%(lineno)d %(message)s"
         logging.basicConfig(format=log_fmt, level=logging.INFO)
 
-        database_path = get_database_path(self.game_version)
+        database_path = get_database_path()
         image_folder = get_image_folder(self.image_folder, self.filepath)
 
         self.import_wismhd(